Made the Aboutus page(not responsive)
